+++ casanovo/config.py
@@ -58,6 +58,7 @@ class Config:
         train_batch_size=int,
         predict_batch_size=int,
         n_beams=int,
+        top_match=int,
         max_epochs=int,
         num_sanity_val_steps=int,
         train_from_scratch=bool,