+++ tests/unit_tests/test_config.py
@@ -17,10 +17,11 @@ def test_override(tmp_path):
     """Test overriding the default"""
     yml = tmp_path / "test.yml"
     with yml.open("w+") as f_out:
-        f_out.write("random_seed: 42")
+        f_out.write("random_seed: 42\ntop_match: 3")
 
     config = Config(yml)
     assert config.random_seed == 42
     assert config["random_seed"] == 42
     assert not config.no_gpu
+    assert config.top_match == 3
     assert config.file == str(yml)