+++ tests/unit_tests/test_unit.py
@@ -1,9 +1,9 @@
+import collections
 import heapq
 import os
 import platform
 import shutil
 import tempfile
-import warnings
 
 import einops
 import github
@@ -159,7 +159,7 @@ def test_beam_search_decode():
     scores = einops.rearrange(scores, "B L V S -> (B S) L V")
     tokens = torch.zeros(batch * beam, length, dtype=torch.int64)
     # Create cache for decoded beams.
-    pred_cache = {i: [] for i in range(batch)}
+    pred_cache = collections.OrderedDict((i, []) for i in range(batch))
 
     # Ground truth peptide is "PEPK".
     true_peptide = "PEPK"
@@ -216,7 +216,7 @@ def test_beam_search_decode():
 
     # Test _get_top_peptide().
     # Return the candidate peptide with the highest score
-    test_cache = {i: [] for i in range(batch)}
+    test_cache = collections.OrderedDict((i, []) for i in range(batch))
     heapq.heappush(
         test_cache[0], (0.93, 4 * [0.93], torch.tensor([4, 14, 4, 19]))
     )
@@ -224,14 +224,19 @@ def test_beam_search_decode():
         test_cache[0], (0.95, 4 * [0.95], torch.tensor([4, 14, 4, 13]))
     )
     heapq.heappush(
-        test_cache[0], (0.94, 4 * [0.94], torch.tensor([4, 14, 4, 14]))
+        test_cache[0], (0.94, 4 * [0.94], torch.tensor([4, 14, 4, 4]))
     )
 
-    assert model._get_top_peptide(test_cache)[0][-1] == "PEPK"
+    assert list(model._get_top_peptide(test_cache))[0][0][-1] == "PEPK"
     # Test that an empty predictions is returned when no beams have been
     # finished.
-    empty_cache = {i: [] for i in range(batch)}
-    assert model._get_top_peptide(empty_cache)[0][-1] == ""
+    empty_cache = collections.OrderedDict((i, []) for i in range(batch))
+    assert len(list(model._get_top_peptide(empty_cache))[0]) == 0
+    # Test multiple PSM per spectrum and if it's highest scoring peptides
+    model.top_match = 2
+    assert set(
+        [pep[-1] for pep in list(model._get_top_peptide(test_cache))[0]]
+    ) == {"PEPK", "PEPP"}
 
     # Test _get_topk_beams().
     # Set scores to proceed generating the unfinished beam.
@@ -263,7 +268,7 @@ def test_beam_search_decode():
     # 1 spectrum with 5 peaks (2 values: m/z and intensity).
     spectra = torch.zeros(1, 5, 2)
     precursors = torch.tensor([[469.25364, 2.0, 235.63410]])
-    assert model.beam_search_decode(spectra, precursors)[0][-1] == ""
+    assert len(list(model.beam_search_decode(spectra, precursors))[0]) == 0
     model.max_length = 100
 
     # Re-initialize scores and tokens to further test caching functionality.
@@ -280,7 +285,7 @@ def test_beam_search_decode():
     scores[:, i, :] = 1
     scores[j, i, tokens[j, i]] = s
 
-    pred_cache = {i: [] for i in range(batch)}
+    pred_cache = collections.OrderedDict((i, []) for i in range(batch))
     finished_beams = torch.tensor([True, True, True, True])
     beam_fits_precursor = torch.BoolTensor([False, True, True, False])
 
@@ -309,7 +314,7 @@ def test_beam_search_decode():
     scores = einops.rearrange(scores, "B L V S -> (B S) L V")
     tokens = torch.zeros(batch * beam, length, dtype=torch.int64)
 
-    pred_cache = {i: [] for i in range(batch)}
+    pred_cache = collections.OrderedDict((i, []) for i in range(batch))
 
     # Ground truth peptide is "PEPK".
     true_peptide = "PEPK$"
@@ -338,6 +343,7 @@ def test_beam_search_decode():
     # Test _finish_beams() for tokens with a negative mass.
     model = Spec2Pep(n_beams=2, residues="massivekb")
     beam = model.n_beams  # S
+    aa2idx = model.decoder._aa2idx
     step = 1
 
     # Ground truth peptide is "-17.027GK".
@@ -355,6 +361,38 @@ def test_beam_search_decode():
     assert torch.equal(finished_beams, torch.tensor([False, True]))
     assert torch.equal(beam_fits_precursor, torch.tensor([False, False]))
     assert torch.equal(discarded_beams, torch.tensor([False, False]))
+
+    # Test _finish_beams() for multiple/internal N-mods and dummy predictions.
+    model = Spec2Pep(n_beams=3, residues="massivekb", min_peptide_len=3)
+    beam = model.n_beams  # S
+    model.decoder.reverse = True
+    aa2idx = model.decoder._aa2idx
+    step = 4
+
+    # Ground truth peptide is irrelevant for this test.
+    precursors = torch.tensor([1861.0044, 2.0, 940.5750]).repeat(
+        beam * batch, 1
+    )
+    tokens = torch.zeros(batch * beam, length, dtype=torch.int64)
+    # Reverse decoding
+    for i, peptide in enumerate(
+        [
+            ["K", "A", "A", "A", "+43.006-17.027"],
+            ["K", "A", "A", "+42.011", "A"],
+            ["K", "A", "A", "+43.006", "+42.011"],
+        ]
+    ):
+        tokens[i, : step + 1] = torch.tensor([aa2idx[aa] for aa in peptide])
+
+    # Test _finish_beams(). All should be discarded
+    finished_beams, beam_fits_precursor, discarded_beams = model._finish_beams(
+        tokens, precursors, step
+    )
+    assert torch.equal(finished_beams, torch.tensor([False, False, False]))
+    assert torch.equal(
+        beam_fits_precursor, torch.tensor([False, False, False])
+    )
+    assert torch.equal(discarded_beams, torch.tensor([False, True, True]))
 
 
 def test_eval_metrics():